[slapd] Deal with attribute name changes in OpenLDAP 2.5+
A couple of attributes changed names in OpenLDAP 2.5, this patch
makes the slapd role deal with both names.

Also the replication example needed updating. Since those values
will be in people's Ansible inventories, we can't fix it but we
can document the change. Maybe something should be added to the
DebOps release notes as well?

With this patch, I think the support for OpenLDAP 2.5 is finished,
so this PR would:

Close: #2244
Close: #2243