Restore extended configuration. Collect static files
I realized that extended configuration is useful, but the previous
because of the previous naming, I thought EteSync removed that feature;
however, it is actually introduced with `etebase_server_settings` which
is more clear.

I also needed to introduce a way to run `collectstatic` which is almost
similar to migration task with little less requirement.